+++ dhall/dhall.cabal
@@ -361,6 +361,7 @@ Library
         Dhall.Pretty.Internal
         Dhall.Syntax
         Dhall.Syntax.Binding
+        Dhall.Syntax.Chunks
         Dhall.Syntax.Const
         Dhall.Syntax.Expr
         Dhall.Syntax.Import
@@ -372,11 +373,11 @@ Library
         Dhall.Syntax.Instances.Functor
         Dhall.Syntax.Instances.Lift
         Dhall.Syntax.Instances.Monad
-        Dhall.Syntax.Instances.Monoid
+        -- Dhall.Syntax.Instances.Monoid
         Dhall.Syntax.Instances.NFData
         Dhall.Syntax.Instances.Ord
         Dhall.Syntax.Instances.Pretty
-        Dhall.Syntax.Instances.Semigroup
+        --Dhall.Syntax.Instances.Semigroup
         Dhall.Syntax.Instances.Show
         Dhall.Syntax.Instances.Traversable
         Dhall.Syntax.MultiLet
