+++ dhall/src/Dhall/Core.hs
@@ -92,7 +92,7 @@ import Prettyprinter          (Pretty)
 
 import qualified Control.Exception
 import qualified Data.Text
-import qualified Dhall.Eval         as Eval
+import qualified Dhall.Eval        as Eval
 
 -- | Pretty-print a value
 pretty :: Pretty a => a -> Text
