+++ dhall/src/Dhall/Import.hs
@@ -235,8 +235,8 @@ import qualified System.AtomicWrite.Writer.ByteString.Binary as AtomicWrite.Bina
 import qualified System.Directory                            as Directory
 import qualified System.Environment
 import qualified System.FilePath                             as FilePath
-import qualified System.IO
 import qualified System.Info
+import qualified System.IO
 import qualified Text.Megaparsec
 import qualified Text.Parser.Combinators
 import qualified Text.Parser.Token
