+++ dhall/src/Dhall/Syntax.hs
@@ -3,6 +3,7 @@ module Dhall.Syntax
     ) where
 
 import Dhall.Syntax.Binding               as Export
+import Dhall.Syntax.Chunks                as Export
 import Dhall.Syntax.Const                 as Export
 import Dhall.Syntax.Expr                  as Export
 import Dhall.Syntax.Import                as Export
@@ -14,11 +15,9 @@ import Dhall.Syntax.Instances.Foldable    as Export ()
 import Dhall.Syntax.Instances.Functor     as Export ()
 import Dhall.Syntax.Instances.Lift        as Export ()
 import Dhall.Syntax.Instances.Monad       as Export ()
-import Dhall.Syntax.Instances.Monoid      as Export ()
 import Dhall.Syntax.Instances.NFData      as Export ()
 import Dhall.Syntax.Instances.Ord         as Export ()
 import Dhall.Syntax.Instances.Pretty      as Export
-import Dhall.Syntax.Instances.Semigroup   as Export ()
 import Dhall.Syntax.Instances.Show        as Export ()
 import Dhall.Syntax.Instances.Traversable as Export ()
 import Dhall.Syntax.MultiLet              as Export
