+++ dhall/src/Dhall/Syntax/Chunks.hs-boot
@@ -1,0 +1,3 @@
+module Dhall.Syntax.Chunks where
+
+data Chunks s a
