+++ dhall/src/Dhall/Syntax/Import.hs
@@ -1,4 +1,6 @@
-{-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE DeriveGeneric     #-}
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE RecordWildCards   #-}
 
 {-| This module contains the core syntax types.
 -}
@@ -15,10 +17,11 @@ module Dhall.Syntax.Import (
     , Scheme(..)
     ) where
 
-import Data.Text         (Text)
-import Dhall.Src         (Src (..))
-import Dhall.Syntax.Expr (Expr (..))
-import GHC.Generics      (Generic)
+import Data.Text                      (Text)
+import Dhall.Src                      (Src (..))
+import Dhall.Syntax.Expr              (Expr (..))
+import Dhall.Syntax.Instances.Functor ()
+import GHC.Generics                   (Generic)
 
 import qualified Dhall.Crypto
 
@@ -90,3 +93,47 @@ data Import = Import
     { importHashed :: ImportHashed
     , importMode   :: ImportMode
     } deriving Generic
+
+
+
+
+instance Semigroup Directory where
+    Directory components0 <> Directory components1 =
+        Directory (components1 <> components0)
+
+instance Semigroup File where
+    File directory0 _ <> File directory1 file =
+        File (directory0 <> directory1) file
+
+instance Semigroup ImportType where
+    Local prefix file0 <> Local Here file1 = Local prefix (file0 <> file1)
+
+    Remote (URL { path = path0, ..}) <> Local Here path1 =
+        Remote (URL { path = path0 <> path1, ..})
+
+    Local prefix file0 <> Local Parent file1 =
+        Local prefix (file0 <> parent <> file1)
+
+    Remote (URL { path = path0, .. }) <> Local Parent path1 =
+        Remote (URL { path = path0 <> parent <> path1, .. })
+
+    import0 <> Remote (URL { headers = headers0, .. }) =
+        Remote (URL { headers = headers1, .. })
+      where
+        importHashed0 = Import (ImportHashed Nothing import0) Code
+
+        headers1 = fmap (fmap (importHashed0 <>)) headers0
+
+    _ <> import1 =
+        import1
+
+instance Semigroup ImportHashed where
+    ImportHashed _ importType0 <> ImportHashed hash importType1 =
+        ImportHashed hash (importType0 <> importType1)
+
+instance Semigroup Import where
+    Import importHashed0 _ <> Import importHashed1 code =
+        Import (importHashed0 <> importHashed1) code
+
+parent :: File
+parent = File { directory = Directory { components = [ ".." ] }, file = "" }
