+++ dhall/src/Dhall/Syntax/Instances/Applicative.hs
@@ -4,6 +4,7 @@ module Dhall.Syntax.Instances.Applicative () where
 
 import Dhall.Syntax.Binding
 import Dhall.Syntax.Expr
+import Dhall.Syntax.Instances.Functor ()
 import Dhall.Syntax.Operations
 import Dhall.Syntax.Types
