+++ dhall/src/Dhall/Syntax/Instances/Bifunctor.hs
@@ -2,9 +2,10 @@
 
 module Dhall.Syntax.Instances.Bifunctor () where
 
-import Data.Bifunctor          (Bifunctor (..))
+import Data.Bifunctor                 (Bifunctor (..))
 import Dhall.Syntax.Binding
 import Dhall.Syntax.Expr
+import Dhall.Syntax.Instances.Functor ()
 import Dhall.Syntax.Operations
 import Dhall.Syntax.Types
