+++ dhall/src/Dhall/Syntax/Instances/Data.hs
@@ -5,8 +5,9 @@
 
 module Dhall.Syntax.Instances.Data () where
 
-import Data.Data          (Data)
+import Data.Data            (Data)
 import Dhall.Syntax.Binding
+import Dhall.Syntax.Chunks
 import Dhall.Syntax.Const
 import Dhall.Syntax.Expr
 import Dhall.Syntax.Types
