+++ dhall/src/Dhall/Syntax/Instances/Eq.hs
@@ -4,8 +4,9 @@
 
 module Dhall.Syntax.Instances.Eq () where
 
-import Data.Bits           (xor)
+import Data.Bits            (xor)
 import Dhall.Syntax.Binding
+import Dhall.Syntax.Chunks
 import Dhall.Syntax.Const
 import Dhall.Syntax.Expr
 import Dhall.Syntax.Import
