+++ dhall/src/Dhall/Syntax/Instances/Foldable.hs
@@ -6,6 +6,7 @@
 module Dhall.Syntax.Instances.Foldable () where
 
 import Dhall.Syntax.Binding
+import Dhall.Syntax.Chunks
 import Dhall.Syntax.Expr
 import Dhall.Syntax.Types
