+++ dhall/src/Dhall/Syntax/Instances/Functor.hs
@@ -6,6 +6,7 @@
 module Dhall.Syntax.Instances.Functor () where
 
 import                Dhall.Syntax.Binding
+import                Dhall.Syntax.Chunks
 import                Dhall.Syntax.Expr
 import {-# SOURCE #-} Dhall.Syntax.Operations (unsafeSubExpressions)
 import                Dhall.Syntax.Types
