+++ dhall/src/Dhall/Syntax/Instances/Lift.hs
@@ -6,6 +6,7 @@
 module Dhall.Syntax.Instances.Lift () where
 
 import Dhall.Syntax.Binding
+import Dhall.Syntax.Chunks
 import Dhall.Syntax.Const
 import Dhall.Syntax.Expr
 import Dhall.Syntax.Types
