+++ dhall/src/Dhall/Syntax/Instances/NFData.hs
@@ -2,8 +2,9 @@
 
 module Dhall.Syntax.Instances.NFData () where
 
-import Control.DeepSeq     (NFData)
+import Control.DeepSeq      (NFData)
 import Dhall.Syntax.Binding
+import Dhall.Syntax.Chunks
 import Dhall.Syntax.Const
 import Dhall.Syntax.Expr
 import Dhall.Syntax.Import
