+++ dhall/src/Dhall/Syntax/Types.hs-boot
@@ -2,8 +2,6 @@ module Dhall.Syntax.Types where
 
 data DhallDouble
 
-data Chunks s a
-
 data PreferAnnotation s a
 
 data RecordField s a
